backend/requirements.txt
httpx==0.28.0
python-dotenv==1.0.1
boto3==1.35.0
pytest==8.3.4
pytest-asyncio==0.24.0
anyio==4.6.2
